+++ boostx/client/example_test.go
@@ -1,15 +1,32 @@
 package client_test
 
 import (
+	"crypto/ecdsa"
+	"crypto/elliptic"
+	"crypto/rand"
 	"fmt"
+	"log"
 	"net/http"
 	"time"
 
 	"github.com/Odds66/boostx-partner-sdk-golang/boostx/client"
+	"github.com/Odds66/boostx-partner-sdk-golang/boostx/keys"
 )
 
+func exampleKeyStore() *keys.StaticPrivateKeyStore {
+	key, err := ecdsa.GenerateKey(elliptic.P256(), rand.Reader)
+	if err != nil {
+		log.Fatal(err)
+	}
+	ks, err := keys.NewStaticPrivateKeyStore(key)
+	if err != nil {
+		log.Fatal(err)
+	}
+	return ks
+}
+
 func ExampleNew() {
-	c := client.New()
+	c := client.New(exampleKeyStore())
 
 	fmt.Println("client created:", c != nil)
 	// Output: client created: true
@@ -17,6 +34,7 @@ func ExampleNew() {
 
 func ExampleNew_withBaseURL() {
 	c := client.New(
+		exampleKeyStore(),
 		client.WithBaseURL("https://custom-api.example.com"),
 	)
 
@@ -34,6 +52,7 @@ func ExampleNew_withHTTPClient() {
 	}
 
 	c := client.New(
+		exampleKeyStore(),
 		client.WithHTTPClient(httpClient),
 	)